+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Corona BE
 
 #### unreleased
+
+- Install cors: `pip install -U flask-cors`